+++ packages/core/src/customer/create-customer-strategy-registry.ts
@@ -31,9 +31,7 @@ import {
     GooglePayWorldpayAccessInitializer,
 } from '../payment/strategies/googlepay';
 import { MasterpassScriptLoader } from '../payment/strategies/masterpass';
-import { StripeScriptLoader } from '../payment/strategies/stripe-upe';
 import { RemoteCheckoutActionCreator, RemoteCheckoutRequestSender } from '../remote-checkout';
-import { ConsignmentActionCreator, ConsignmentRequestSender } from '../shipping';
 import {
     createSpamProtection,
     SpamProtectionActionCreator,
@@ -55,7 +53,6 @@ import { DefaultCustomerStrategy } from './strategies/default';
 import { GooglePayCustomerStrategy } from './strategies/googlepay';
 import { MasterpassCustomerStrategy } from './strategies/masterpass';
 import { SquareCustomerStrategy } from './strategies/square';
-import { StripeUPECustomerStrategy } from './strategies/stripe-upe';
 
 export default function createCustomerStrategyRegistry(
     store: CheckoutStore,
@@ -291,21 +288,6 @@ export default function createCustomerStrategyRegistry(
                 remoteCheckoutActionCreator,
                 createGooglePayPaymentProcessor(store, new GooglePayWorldpayAccessInitializer()),
                 formPoster,
-            ),
-    );
-
-    registry.register(
-        'stripeupe',
-        () =>
-            new StripeUPECustomerStrategy(
-                store,
-                new StripeScriptLoader(scriptLoader),
-                customerActionCreator,
-                paymentMethodActionCreator,
-                new ConsignmentActionCreator(
-                    new ConsignmentRequestSender(requestSender),
-                    new CheckoutRequestSender(requestSender),
-                ),
             ),
     );
