--- packages/core/src/customer/strategies/stripe-upe/index.ts
@@ -1,2 +1,0 @@
-export { default as StripeUPECustomerStrategy } from './stripe-upe-customer-strategy';
-export { default as StripeUPECustomerInitializeOptions } from './stripeupe-customer-initialize-options';
