+++ packages/core/src/payment/create-payment-strategy-registry.spec.ts
@@ -41,7 +41,6 @@ import { PPSDKStrategy } from './strategies/ppsdk';
 import { QuadpayPaymentStrategy } from './strategies/quadpay';
 import { SagePayPaymentStrategy } from './strategies/sage-pay';
 import { SquarePaymentStrategy } from './strategies/square';
-import { StripeUPEPaymentStrategy } from './strategies/stripe-upe';
 import { WepayPaymentStrategy } from './strategies/wepay';
 import { ZipPaymentStrategy } from './strategies/zip';
 
@@ -287,12 +286,6 @@ describe('CreatePaymentStrategyRegistry', () => {
         const paymentStrategy = registry.get(PaymentStrategyType.STRIPE_UPE_GOOGLE_PAY);
 
         expect(paymentStrategy).toBeInstanceOf(GooglePayPaymentStrategy);
-    });
-
-    it('can instantiate stripe-upe', () => {
-        const paymentStrategy = registry.get(PaymentStrategyType.STRIPE_UPE);
-
-        expect(paymentStrategy).toBeInstanceOf(StripeUPEPaymentStrategy);
     });
 
     it('can instantiate wepay', () => {
