--- packages/core/src/payment/strategies/stripe-upe/index.ts
@@ -1,5 +1,0 @@
-export * from './stripe-upe';
-
-export { default as StripeScriptLoader } from './stripe-upe-script-loader';
-export { default as StripeUPEPaymentStrategy } from './stripe-upe-payment-strategy';
-export { default as StripeUPEPaymentInitializeOptions } from './stripe-upe-initialize-options';
