+++ packages/payment-integration-api/src/customer/customer.ts
@@ -18,6 +18,7 @@ export default interface Customer {
      * Note: You need to enable "Prompt existing accounts to sign in" in your Checkout Settings.
      */
     shouldEncourageSignIn: boolean;
+    isStripeLinkAuthenticated?: boolean;
     customerGroup?: CustomerGroup;
 }
