+++ packages/stripe-integration/src/index.ts
@@ -1,8 +1,23 @@
 export { default as createStripeV3PaymentStrategy } from './stripev3/create-stripev3-payment-strategy';
+export { default as createStripeUPEPaymentStrategy } from './stripe-upe/create-stripe-upe-payment-strategy';
+export { default as createStripeUPECustomerStrategy } from './stripe-upe/create-stripe-upe-customer-strategy';
 
 export { default as StripeScriptLoader } from './stripev3/stripev3-script-loader';
+export { default as StripeUPEScriptLoader } from './stripe-upe/stripe-upe-script-loader';
+
 export { default as StripeV3PaymentStrategy } from './stripev3/stripev3-payment-strategy';
+export { default as StripeUPEPaymentStrategy } from './stripe-upe/stripe-upe-payment-strategy';
+export { default as StripeUPECustomerStrategy } from './stripe-upe/stripe-upe-customer-strategy';
+
 export {
     default as StripeV3PaymentInitializeOption,
     WithStripeV3PaymentInitializeOptions,
 } from './stripev3/stripev3-initialize-options';
+export {
+    default as StripeUPEPaymentInitializeOption,
+    WithStripeUPEPaymentInitializeOptions,
+} from './stripe-upe/stripe-upe-initialize-options';
+export {
+    default as StripeUPECustomerInitializeOption,
+    WithStripeUPECustomerInitializeOptions,
+} from './stripe-upe/stripeupe-customer-initialize-options';
