+++ packages/stripe-integration/src/stripe-upe/format-locale.ts
