+++ packages/stripe-integration/src/stripe-upe/stripe-upe-customer.mock.ts
@@ -1,4 +1,8 @@
-import { StripeElement, StripeUPEClient } from '../../../payment/strategies/stripe-upe';
+// eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
+import {
+    StripeElement,
+    StripeUPEClient,
+} from '../../../../../stripe-integration/src/stripe-upe/stripe-upe';
 import { CustomerInitializeOptions } from '../../customer-request-options';
 
 export function getCustomerStripeUPEJsMock(returnElement?: StripeElement): StripeUPEClient {
