+++ packages/stripe-integration/src/stripe-upe/stripe-upe-initialize-options.ts
@@ -36,3 +36,7 @@ export default interface StripeUPEPaymentInitializeOptions {
 
     render(): void;
 }
+
+export interface WithStripeUPEPaymentInitializeOptions {
+    stripeupe?: StripeUPEPaymentInitializeOptions;
+}
