+++ packages/stripe-integration/src/stripe-upe/stripe-upe-supported-locales.ts
