+++ packages/stripe-integration/src/stripe-upe/stripeupe-customer-initialize-options.ts
@@ -38,3 +38,11 @@ export default interface StripeUPECustomerInitializeOptions {
           }
         | undefined;
 }
+
+export interface WithStripeUPECustomerInitializeOptions {
+    /**
+     * The options that are required to initialize the customer step of checkout
+     * when using StripeUPE.
+     */
+    stripeupe?: StripeUPECustomerInitializeOptions;
+}
